+++ platformio/downloader.py
@@ -3,15 +3,16 @@
 
 from email.utils import parsedate_tz
 from math import ceil
+from os import environ
 from os.path import getsize, join
 from time import mktime
 
-from click import progressbar
-from requests import get
+import click
+import requests
 
+from platformio import util
 from platformio.exception import (FDSHASumMismatch, FDSizeMismatch,
                                   FDUnrecognizedStatusCode)
-from platformio.util import change_filemtime, exec_command
 
 
 class FileDownloader(object):
@@ -27,7 +28,8 @@ class FileDownloader(object):
             self.set_destination(join(dest_dir, self._fname))
         self._progressbar = None
 
-        self._request = get(url, stream=True)
+        self._request = requests.get(url, stream=True,
+                                     headers=util.get_request_defheaders())
         if self._request.status_code != 200:
             raise FDUnrecognizedStatusCode(self._request.status_code, url)
 
@@ -48,9 +50,14 @@ class FileDownloader(object):
         f = open(self._destination, "wb")
         chunks = int(ceil(self.get_size() / float(self.CHUNK_SIZE)))
 
-        with progressbar(length=chunks, label="Downloading") as pb:
-            for _ in pb:
+        if environ.get("CI") == "true":
+            click.echo("Downloading...")
+            for _ in range(0, chunks):
                 f.write(next(itercontent))
+        else:
+            with click.progressbar(length=chunks, label="Downloading") as pb:
+                for _ in pb:
+                    f.write(next(itercontent))
         f.close()
         self._request.close()
 
@@ -66,11 +73,12 @@ class FileDownloader(object):
 
         dlsha1 = None
         try:
-            result = exec_command(["sha1sum", self._destination])
+            result = util.exec_command(["sha1sum", self._destination])
             dlsha1 = result['out']
         except OSError:
             try:
-                result = exec_command(["shasum", "-a", "1", self._destination])
+                result = util.exec_command(
+                    ["shasum", "-a", "1", self._destination])
                 dlsha1 = result['out']
             except OSError:
                 pass
@@ -83,7 +91,7 @@ class FileDownloader(object):
     def _preserve_filemtime(self, lmdate):
         timedata = parsedate_tz(lmdate)
         lmtime = mktime(timedata[:9])
-        change_filemtime(self._destination, lmtime)
+        util.change_filemtime(self._destination, lmtime)
 
     def __del__(self):
         self._request.close()