+++ platformio/ide/__init__.py
@@ -1,0 +1,2 @@
+# Copyright (C) Ivan Kravets <me@ikravets.com>
+# See LICENSE for details.