+++ scripts/appveyor/install.ps1
@@ -74,7 +74,7 @@ function InstallPip ($python_home) {
 
 function InstallPackage ($python_home, $pkg) {
     $pip_path = $python_home + "/Scripts/pip.exe"
-    & $pip_path install $pkg
+    & $pip_path install -U $pkg
 }
 
 function main () {