+++ HISTORY.rst
@@ -1,6 +1,51 @@
 Release History
 ===============
 
+2.0.0 (2015-05-22)
+------------------
+
+* PlatformIO as :ref:`ci` (CI) tool for embedded projects
+  (`issue #108 <https://github.com/platformio/platformio/issues/108>`_)
+* Initialise PlatformIO project for the specified IDE
+  (`issue #151 <https://github.com/platformio/platformio/issues/151>`_)
+* PlatformIO CLI 2.0: "platform" related commands have been
+  moved to ``platformio platforms`` subcommand
+  (`issue #158 <https://github.com/platformio/platformio/issues/158>`_)
+* Created `PlatformIO gitter.im <https://gitter.im/platformio/platformio>`_ room
+  (`issue #174 <https://github.com/platformio/platformio/issues/174>`_)
+* Global ``-f, --force`` option which will force to accept any
+  confirmation prompts
+  (`issue #152 <https://github.com/platformio/platformio/issues/152>`_)
+* Run project with `platformio run --project-dir <http://docs.platformio.org/en/latest/userguide/cmd_run.html#cmdoption--project-dir>`_ option without changing the current working
+  directory
+  (`issue #192 <https://github.com/platformio/platformio/issues/192>`_)
+* Control verbosity of `platformio run <http://docs.platformio.org/en/latest/userguide/cmd_run.html#cmdoption-platformio-run-v>`_ command via ``-v/--verbose`` option
+* Add library dependencies for build environment using
+  `install_libs <http://docs.platformio.org/en/latest/projectconf.html#install-libs>`_
+  option in ``platformio.ini``
+  (`issue #134 <https://github.com/platformio/platformio/issues/134>`_)
+* Specify libraries which are compatible with build environment using
+  `use_libs <http://docs.platformio.org/en/latest/projectconf.html#use-libs>`_
+  option in ``platformio.ini``
+  (`issue #148 <https://github.com/platformio/platformio/issues/148>`_)
+* Add more boards to PlatformIO project with
+  `platformio init --board <http://docs.platformio.org/en/latest/userguide/cmd_init.html#cmdoption--board>`__
+  command
+  (`issue #167 <https://github.com/platformio/platformio/issues/167>`_)
+* Choose which library to update
+  (`issue #168 <https://github.com/platformio/platformio/issues/168>`_)
+* Specify `platformio init --env-prefix <http://docs.platformio.org/en/latest/userguide/cmd_init.html#cmdoption--env-prefix>`__ when initialise/update project
+  (`issue #182 <https://github.com/platformio/platformio/issues/182>`_)
+* Added new Armstrap boards
+  (`issue #204 <https://github.com/platformio/platformio/issues/204>`_)
+* Updated SDK for `espressif <http://docs.platformio.org/en/latest/platforms/espressif.html>`__
+  development platform to v1.1
+  (`issue #179 <https://github.com/platformio/platformio/issues/179>`_)
+* Disabled automatic updates by default for platforms, packages and libraries
+  (`issue #171 <https://github.com/platformio/platformio/issues/171>`_)
+* Fixed bug with creating copies of source files
+  (`issue #177 <https://github.com/platformio/platformio/issues/177>`_)
+
 1.5.0 (2015-05-15)
 ------------------
 
@@ -42,8 +87,9 @@ Release History
   1.6.3 version (`issue #156 <https://github.com/platformio/platformio/issues/156>`_)
 * Upgraded `Energia Framework <http://docs.platformio.org/en/latest/frameworks/energia.html>`__ to
   0101E0015 version (`issue #146 <https://github.com/platformio/platformio/issues/146>`_)
-* Upgraded `Arduino Framework with Teensy Core <http://docs.platformio.org/en/latest/frameworks/arduino.html>`_ to
-  1.22 version (`issue #162 <https://github.com/platformio/platformio/issues/162>`_,
+* Upgraded `Arduino Framework with Teensy Core <http://docs.platformio.org/en/latest/frameworks/arduino.html>`_
+  to 1.22 version
+  (`issue #162 <https://github.com/platformio/platformio/issues/162>`_,
   `issue #170 <https://github.com/platformio/platformio/issues/170>`_)
 * Fixed exceptions with PlatformIO auto-updates when Internet connection isn't
   active